Move `PatchMerging` into utils and use common `PatchEmbed`.
