+++ mmcls/models/utils/__init__.py
@@ -1,6 +1,6 @@
 from .augment.augments import Augments
 from .channel_shuffle import channel_shuffle
-from .embed import HybridEmbed, PatchEmbed
+from .embed import HybridEmbed, PatchEmbed, PatchMerging
 from .helpers import to_2tuple, to_3tuple, to_4tuple, to_ntuple
 from .inverted_residual import InvertedResidual
 from .make_divisible import make_divisible
@@ -8,6 +8,6 @@ from .se_layer import SELayer
 
 __all__ = [
     'channel_shuffle', 'make_divisible', 'InvertedResidual', 'SELayer',
-    'to_ntuple', 'to_2tuple', 'to_3tuple', 'to_4tuple', 'Augments',
-    'HybridEmbed', 'PatchEmbed'
+    'to_ntuple', 'to_2tuple', 'to_3tuple', 'to_4tuple', 'PatchEmbed',
+    'PatchMerging', 'HybridEmbed', 'Augments'
 ]
